How-to-use-Weaviate/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
Authlib==1.3.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
cryptography==44.0.2
datasets==3.5.0
dill==0.3.8
filelock==3.18.0
frozenlist==1.6.0
fsspec==2024.12.0
grpcio==1.71.0
grpcio-health-checking==1.71.0
grpcio-tools==1.71.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
multidict==6.4.3
multiprocess==0.70.16
numpy==2.2.4
packaging==25.0
pandas==2.2.3
propcache==0.3.1
protobuf==5.29.4
pyarrow==19.0.1
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
setuptools==79.0.1
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
validators==0.34.0
weaviate-client==4.13.2
xxhash==3.5.0
yarl==1.20.0
